Check if vhost already exists Clean package.json

+++ src/commands/namespaces/project.js
@@ -6,13 +6,15 @@ var queue = require('../../utils/queue');
 var colors = require('../../utils/consoleColors');
 
 function checkoutBranch(branch){
+    console.log(colors.green + "\tChecking out branch " + colors.white + branch + colors.reset);
     exec('git checkout ' + branch, queue.add('project', function(error, stdout, stderr){
         if (error) {
             console.log(colors.red + "\tError while checkout git" + colors.reset);
             console.log(stderr);
             return;
         }
-
+        
+        console.log(colors.green + "\tPulling branch " + colors.white + branch + colors.reset);
         exec('git pull origin ' + branch, queue.add('project', function(error, stdout, stderr){
             if (error) {
                 console.log(colors.red + "\tError while pulling branch :" + colors.reset);
@@ -23,21 +25,34 @@ function checkoutBranch(branch){
 }
 
 function createVhost(host, port, root){
-    var vhost = "<VirtualHost *:" + port + ">\n\tServerName " + host + "\n\tDocumentRoot \"" + root + "\"\n</VirtualHost>", error;
+    var vhost = "<VirtualHost *:" + port + ">\n\tServerName " + host + "\n\tDocumentRoot \"" + root + "\"\n</VirtualHost>", content, error
 
     console.log(colors.green + "\tWriting virtual host" + colors.reset);
 
     if (fs.existsSync("/etc/apache2/extra") && fs.existsSync("/etc/apache2/extra/httpd-vhosts.conf")) {
+        content = fs.readFileSync('/etc/apache2/extra/httpd-vhosts.conf');
+
+        if (content.toString().indexOf(vhost)) {
+            console.log(colors.white + "\tVirtual host already exists" + colors.reset);
+            return;
+        }
+
         try{
             fs.appendFileSync("/etc/apache2/extra/httpd-vhosts.conf", "\n" + vhost);
             console.log(colors.green + "\tYou can restart apache to enable your new virtual host" + colors.reset);
         } catch (error) {
-            console.log(colors.red + "\tError while writing virtual host, try using sudo" + colors.reset);
+            console.log(colors.red + "\tCan't write on virtual host file, try using sudo" + colors.reset);
         }
+
         return;
     }
 
     if (fs.existsSync("/etc/apache2/sites-available/")) {
+        if(fs.existsSync("/etc/apache2/sites-available/" + host + ".conf")){
+            console.log(colors.white + "\tVirtual host already exists" + colors.reset);
+            return;
+        }
+
         fs.writeFileSync("/etc/apache2/sites-available/" + host + ".conf", vhost);
 
         exec("ln -s /etc/apache2/sites-available/" + host + ".conf /etc/apache2/sites-enabled/" + host + ".conf", queue.add('project', function(error, stdout, stderr){
@@ -52,12 +67,12 @@ function createVhost(host, port, root){
 
 module.exports = function(args, rootDir, next){
 
-    if (args.branch) {
-        queue.add('project', checkoutBranch)(args.branch);
-    }
-
     if (args.host) {
         queue.add('project', createVhost)(args.host, args.port || 80, rootDir + '/' + (args.root || ''));
+    }
+
+    if (args.branch) {
+        queue.add('project', checkoutBranch)(args.branch);
     }
 
     queue.done('project', next);
